Changelog 10/04/2023
Added a history of the last four requests.
Clicking on a request populates the search bar with the query used to retrieve information about the account, and displays all information associated with that account.
Renamed some fields and methods in the ProfileDataFetcherViewModel.
Added a new field, "Request", to SteamProfile to store the user's query used to retrieve data about the account.